requirements.txt
# FastAPI
fastapi
uvicorn[standard]

# Database
sqlalchemy
psycopg2-binary  # Driver cho PostgreSQL
asyncpg          # Driver bất đồng bộ cho PostgreSQL
pgvector         # Client library cho pgvector

# LangChain & Google Gemini
langchain
langgraph
langchain-google-genai
langchain-community # Cần cho một số integration
google-generativeai

# Khác
pydantic-settings  # Để quản lý config từ env vars
python-dotenv